[specs] Confirm homepage redirect before proceeding to logs [DEV-39]
Hypothesis: `visit(logs_path)` occurs quickly enough to cancel the
request that is being made to log in the user.

Hoped for solution: confirm the path and page content of the home page
(to which the user will be redirected after being logged in). _Then_
visit the logs path (hopefully now with the assurance that the login
request went to completion rather than being cancelled).