[specs] Confirm homepage redirect before proceeding to logs [DEV-39] (#4832)
Hypothesis: `visit(logs_path)` sometimes occurs quickly enough to cancel
the browser request that is being made to log in the user.

Hoped for solution: confirm the path and page content of the home page
(to which the user will be redirected after logging in). _Then_ visit
the logs path (hopefully now with the assurance that the login request
went to completion, rather than being cancelled).