Implement zero-downtime, rolling deploys [DEV-65] (#4835)
https://www.tines.com/blog/simple-zero-downtime-deploys-with-nginx-and-docker-compose/

^ Really great article. I am amazed at how relatively simple this is.

One difference in my experience vs what's described in that article is
that I don't seem to ever be in a phase where there is a mix of requests
going to old and new. For me, there's just a sudden shift from the old
to the new. Once the new is ready, then it's almost immediate that we
switch over to exclusively routing requests to the new container.