Add nginx and web to 'nondefault' Docker Compose profile [DEV-65]
This will avoid these services being restarted when we call `docker
compose up --detach --remove-orphans` in `deploy.sh`. (If these services
are restarted by that command, then we get a significant chunk of
downtime, which we don't want. Instead, we want to manage the rollout of
`web` only via the `roll-out-web.sh` script.)

It still happens sometimes that some requests get caught in the
transition, and I believe that they hang for the duration of the NGINX
`proxy_connect_timeout`. I am therefore going to lower that timeout, so
that these requests fail more quickly. I'm also lowering the
`proxy_send_timeout` and `proxy_read_timeout` while I'm at it.