Delete app.json (#4850)
This was a Dokku thing. We currently have healthchecking sort of built
into `bin/server/roll-out-web.sh`, although that doesn't currently check
for the text "David Runger"; it just checks for a 200 status code. I
think it's fairly unlikely that the Rails server will respond with a 200
status code and not with the text "David Runger", though.