[runger-config/spec-commands] Don't set TERM=dumb (#4860)
I thought that this is needed to get tests to pass locally when run via
`spec`, but instead I am going to modify `z.rb` not to load string
monkeypatches except in development mode, which makes `TERM=dumb`
unnecessary (if it ever even helped at all).

This undoes part of the changes recently made in #4859.