[spec/logs] Sleep 0.8 seconds before publishing new log entry [LOG-8]
This change aims to fix a flaky spec. See a failure [here][1].

**Hypothesis:** It's possible that the log entry is published via
ActionCable before the ActionCable JavaScript in the browser has had
time to subscribe to the relevant WebSocket.

**Solution:** Modify application code (sad) to add a globally accessible
marker (`window.davidrunger.connectedToLogEntriesChannel`) when the
ActionCable subscription has been created.

I actually saw this flake while working on this branch, after having
added the `wait_for` to wait for
`window.davidrunger.connectedToLogEntriesChannel` to be true, so
apparently that alone is not sufficient. Therefore, let's also add a 0.5
second sleep after the ActionCable consumer has connected.

Hopefully these two changes together will eliminate the flakiness that
we have seen in this spec.

[1]: https://github.com/davidrunger/david_runger/actions/runs/9737531169/job/26869781563#step:9:319