[specs] Bring back rspec-retry, only for Ferrum::ProcessTimeoutError (#4868)
I don't really know of anything that we can do to avoid these
`Ferrum::ProcessTimeoutError` exceptions that occur from time to time
(not only in specs, but also in our prerenderer). When they do occur,
let's retry (and log), rather than failing the build.

We had removed `rspec-retry` in #2417 (though noting that we might want
to bring it back at some point).