Stop adding .env.{development,production}.local files to docker image (#4879)
As of this change, it's no longer necessary to add the
`.env.{development,production}.local` files to the image; it's
sufficient for one or both of these env files just to exist on the host
machine. (In development, we actually need both, so that the correct
DATABASE_URL is defined, for example, though it will be connected to my
local almost-empty "production" Docker database, rather than to the real
production database.)

Prior to this change, we did need these files to be added to the image,
so that they would be available during the Dockerfile build process when
we run `assets:precompile`, so that we'd have access to an AWS access
key and secret, so that we could download the precompiled frontend
assets.

To avoid this need, this change uses an undocumented
`unsigned_operations` AWS SDK option
(https://github.com/aws/aws-sdk-ruby/issues/
1149#issuecomment-2007175332), which makes it so that the AWS SDK
doesn't attempt to sign the request for the publicly accessible build
assets, which makes it so that no AWS access key and/or secret are
needed during the Dockerfile build process (and nor are any other
secrets needed).

**Motivation:** By making it so that no secrets are required in order to
build the image for our Rails app, we'll be able to transition to
building the images via GitHub Actions, rather than on the server
(DEV~71).