[nginx] Gzip 'text/html' and 'text/javascript' responses
We already have `application/javascript` in the list, but our JavaScript
files are being served as `text/javascript`, which apparently is also a
valid JavaScript MIME type, so here we add that to the list, as well.

Prior to this, Cloudflare was compressing uncompressed files (of
nontrivial size) with brotli, so for e.g. HTML files I think that we
were uncompressed from our server to Cloudflare, but then compressed
with brotli from Cloudflare to the user. Hopefully compressing the whole
trip from our server with gzip will be faster overall? If it's not, we
might want to disable all gzip compression on our server, and just let
Cloudflare compress everything as brotli.