Revert "Use git_rev action (not 'git rev-parse origin/main')"
This reverts commit 8f1e5b6b5972e5e21540c41e75735537ec7e78ee.

We need the origin/main SHA for the build so that compiled assets can be
downloaded.

It would be nice to tag with the true Git SHA, but the
docker/metadata-action action does what it does, regardless of what
GIT_REV is. And it shouldn't make a difference one way or another when
we run this on `main`.