[enable_yjit] Move server conditional within after_initialize block
This way, I think that, when running RSpec feature tests, we will enable
YJIT for the server that is used by Capybara, which is maybe a good
thing? Another advantage is that it will bring these lines under test
coverage, which they aren't currently (since, when the `bin/rspec`
process boots up (i.e. at the level of the outer conditional) I think
`Rails::Server` is not yet defined).