Revert "Build Docker image on GitHub (not on server) [DEV-71] (#4903)"
This reverts commit 081c2d24eb328095952df9524367b8178a116098.

**Motivation:** `--mount=type=cache` does not work in GitHub Actions
(https://github.com/moby/buildkit/issues/ 1512), which means we cannot
cache gems between builds. We can still cache the overall Docker build
step, so, as long as no gems change, the build is quick. However, a
decent fraction of our PRs do involve gems changing, so this seems worth
accounting for. Let's just try going back to building the images on the
server.

Part of the motivation for building in GitHub Actions rather than on the
server was because we were running out of memory when trying to deploy
to the server and because the swap management process was using tons of
CPU. However, now that I have allotted 2 GB of swap space (we had 0
before) and also increased the swappiness from 0 to currently 25 (and
maybe I will go even higher, especially with this change), I'm
optimistic that the server will now be able to handle building the
images (and also having 2 web instances booted simultaneously when we do
our zero-downtime deployment rollout).