[home/projects] Add emoji picker (#4920)
It's a pretty basic project, but it has the advantage of not being
behind an authentication wall (unlike all of the other mentioned
projects), so I think it's worth listing.

It's also just a single file, which is sort of nice, because I can link
to that file on GitHub (vs my other projects, which are spread across
many files, and so cannot be linked to in this way).