[docker] Don't send development logs to Papertrail
In development, developers can avoid sending logs to Papertrail (and
instead just have logs available locally, via the `local` driver) via
`cp docker-compose.override.yml.example docker-compose.override.yml`.
The `docker-compose.override.yml` is a special file name that is loaded
automatically by Docker.

The reason that we are using a template (and gitignoring the actual
`docker-compose.override.yml` file) is because we _do_ want to track
this file via version control, but we _don't_ want this file to be sent
to the server via git, since on the server we do want logs to be sent to
Papertrail.

Also, instead of using `.env.papertrail.local` to source
`PAPERTRAIL_URL`, we will now use `.env`, which is sourced by Docker
Compose automatically, which makes things a lot easier and which is a
more complete/foolproof solution. For example, even with the code being
deleted in this PR (trying to load `.env.papertrail.local` manually)
doesn't help when e.g. running `docker compose run --rm web echo 'hi'`
on the server; we'd still get warnings like this:

> WARN[0000] The "PAPERTRAIL_URL" variable is not set. Defaulting to a blank string.

Just using `.env` (on the server) should avoid these warnings in all
cases.

Also, in order to avoid warnings in development if a `.env` file with
`PAPERTRAIL_URL` is not present, we also now provide a default value for
`PAPERTRAIL_URL` in `docker-compose.yml`.