[docker-compose] Add health check for certbot service (#4956)
This doesn't really check that certbot is actually doing it's certbot
work or even that it's running at all (which, it usually won't be; it
will just be sleeping). This just checks that `certbot --version` is
available. Better than nothing, though, I guess. And `certbot` was the
only service without a health check; now all services have one.