Use render_schema_json and JSON-schema-generated types more [DEV-99] (#5025)
... a lot more.

I tried to generate a JSON schema for pretty much every JSON response
that isn't just an error(s). In at least a couple places, where the JSON
response isn't actually being used by the frontend, I switched to `head
:ok`, to make it more clear that the JSON isn't currently needed/used on
the frontend.

I'm removing the `schema` keyword arg from the `render_schema_json`
method. I think it makes significantly more sense to have one schema per
action, even if there ends up being duplication between some of the
actions. When the client side code makes a certain request, it makes a
lot more sense for the type used to be specific for that request, rather
than for the type name to randomly refer to some other request, just
because that's how we set it up on the backend, I guess to reduce
duplication.

Overall / stepping back, I'm not super sure whether this is worthwhile.
I can imagine concluding that it's not, and undoing some of this. But,
that being said, I can also imagine that this might possibly
catch/prevent some bugs. I do feel like it fills what I have felt as a
gap in the application's structure; it does feel satisfying to connect
the backend and frontend in this way.

In development mode, the Api::BaseController tries to help out as much
as it can, creating the necessary schema json file if it's missing,
opening it in $EDITOR, and copying the JSON response data into the
clipboard, for easy copy-pasting (after privacy sanitization) into
https://jsonformatter.org/json-to-jsonschema .