[groceries/spec] Ensure that add-item API request succeeds [GROC-19] (#5026)
We'll check for 3 things:
1. No spinner on the page
2. Input value is cleared
3. "Add" button is disabled (since input should be emptied)

I'm decreasing the `sleep` from 1 second to 0.5 seconds, because I think
that these checks will do some of the work that the `sleep` was
previously doing.

Also, add explicit negated matching condition and message for
`have_flash_message` matcher, because (based on my experience with
adding the `have_spinner` matcher) it seems that there is a risk of
hanging indefinitely when testing a negated condition without providing
one or both of these.