[spec/quizzes] Add more page content expectations [QUIZ-4]
It seems like maybe sometimes a response intended for the quiz owner
gets processed in the session of the non-owner quiz participant? (If so,
hopefully that's not something that can happen in real life, but rather
is just a quirk of Capybara's session management.)

This change adds additional expectations about page content after taking
actions in  the browser. Hopefully this will slow things down a bit /
keep things a bit more stable, so that responses are less likely to get
mixed up as we move quickly back and forth between the two sessions.

I ran the test 60 times after making this change, and it flaked with the
error that this change aims to fix (`TurboFrameMissingError`) zero
times, whereas, when I did the same thing previously without this
change, the spec flaked 5 times with `TurboFrameMissingError`. So, I
think that's pretty encouraging.

Other flakiness remains in the spec being modified herein, but I think
that I'll try to fix that in a separate PR.