[groceries] Render toast upon successful check-in [GROC-20]
Also, render an error message if the check-in errors. To support this,
allow specifying a `type` for `vueToast`.

Also, disable the check-in button while the network request is in
flight.

Also, make `:notice` the default expected type for the
`have_flash_message` matcher.

Also, add a new `have_vue_toast` matcher. (I considered
extending/renaming the existing `have_flash_message` matcher, instead,
but decided against it, due to concerns about a combined approach
potentially being complex and/or non-performant.)

Also, confirm in the groceries feature spec that the checked-in item is
reduced to zero needed.