[rollbar] Use 'git' in block variant of ENV.fetch (#5063)
Otherwise, we try to call out to `git`, even if we don't actually need
it (and it might not be available).

I think this will fix complaints such as `fatal: not a git repository
(or any of the parent directories): .git`
(https://github.com/davidrunger/david_runger/actions/runs/10650392835/job/29521750853#step:5:94)
and `sh: 1: git: not found`
(https://github.com/davidrunger/david_runger/actions/runs/10650392835/job/29521750853#step:5:466).

This issue was introduced in
https://github.com/davidrunger/david_runger/pull/5061 .