Use delegated types for log entries [LOG-21] (#5085)
Using delegated types might add some net significant lines of code
(though it's hard to tell for sure from the diff, since a fair amount of
the diff is incidental additional and deletions), and using delegated
types is arguably, in at least some sense, somewhat more complex. In
particular, there is some delicacy and additional complication when
creating or updating records, because now we must update two separate
records, rather than updating just a single record, as we did before.

However, that being said, I do kind of think it's an improvement / the
right way to go. Our old data structure was decently confusing, in its
own right, and I always had the nagging feeling that it is somehow
"wrong". This new approach feels a little more "right" to me. For one
thing, the data structure is DRYer; for example, after cleanup (though
not as of this PR) we won't duplicate the `note` column across multiple
tables.