[specs] Raise error if ApplicationRecord::Base.connection is called
fixes https://app.rollbar.com/a/davidjrunger/fix/item/davidrunger/672

Unfortunately, I think that this requires monkeypatching ActiveRecord.
At least, that's the only solution that I can think of. So, that's what
I'm doing. Fortunately, it's only in test that we need to do it. The
reason is that something about how tests set up their database
connection (which I think then gets used by the application code) causes
the error to not occur in some situations when it would actually occur
in development or production. In the original ActiveRecord code, the
`permanent_connection_checkout` config is only checked if
`pool.permanent_lease?` is truthy. It's falsy in test, though. This
monkeypatch changes the code to check the
`permanent_connection_checkout` config whether `pool.permanent_lease?`
is truthy or falsy.

This also requires forking ActiveAdmin, because it uses
`ActiveRecord::Base.connection` (via `resource_class.connection`) in one
place. I have submitted a PR to upstream my fork:
https://github.com/activeadmin/activeadmin/pull/ 8474 .