[docker-compose] Comment out logging temporarily
We have hit our papertrail log limit for the month (due to an issue
where we were logging a bunch of stack traces). Unfortunately, I think
that this means that papertrail refuses to even connect, resulting in
deployment failures. Comment out this logging temporarily, until the
monthly quote resets on 2024-09-17.