Allow permanent DB connection for database_consistency and immigrant (#5171)
I am broadening the monkeypatch scope from just in test to also include
development, because I want to be able to run these checks in both
development and test without error.

Also, change the containing module from `Runger` (provided (but
shouldn't be?) by the `runger_config` gem?) to `DavidRunger`, which is
defined in `config/application.rb`.