Store JWT secret in ENV var (not Rails credentials) (#5434)
This is somewhat of a follow-up to ef138c27 / #4731 .

Earlier this year, I lost the encryption secret for the development
credentials file. Rather than rebuild a new credentials file (which sort
of demonstrates how credentials can be a bit of a pain), instead, we'll
continue herein migrating from Rails credentials to ENV vars.

I have just set a `JWT_SECRET` env var in production (to the same value
as `Rails.application.credentials.jwt_secret` was), so this should be a
smooth, no-downtime deployment.