[logs] Delete 'unsafeSelectedLog' method from store
The deleted method was causing issues. It seems that, as of pinia 2.2.5,
apparently this getter function gets evaluated, just by virtue of
existing? If I change the name to `unsafeSelectedLogz` (which is called
nowhere), then it still gets evaluated when going to `/logs`.

It's not very difficult to move the `assert`s one level out, so I'm
going to try that and see if it resolves all of the spec failures.