Don't stash initial request data in Redis for HealthChecksController (#5774)
The requests to `/up` were using up a nontrivial amount of storage (~90
MB or so, IIRC) in the application's Redis database. I believe that the
requests were sitting in Redis for their full 21 day TTL. This change
will make it so that we don't store that data in Redis for any period of
time, which should make Redis use notably less memory overall.