[vector] Don't send host/system to Papertrail
Previously, the container id of the vector container was being sent as
the host/system, which is not particularly helpful, and which I think
might have some downsides. When no host/system is provided, I think
Papertrail will just log the sending IP address as the system, which I
think actually works pretty well, so this change should bring up back to
that.