Store logs in Loki and view them via Grafana (via NGINX) [DEV-120]
Docker's built-in json-file logging seems to not work very well, at
least when also using log rotation. For example, `docker compose logs
<service_name>` only shows logs for the current container, even if that
container was just booted up recently. There are other problems. And we
can't currently afford to send all logs (including NGINX) to Papertrail.
So, we don't currently have a good solution that covers NGINX.
Loki/Grafana will provide that.

A side benefit / win-win is that it would also be nice to have Grafana
available for viewing metrics (aggregated/stored via StatsD, Graphite,
and/or whatever), and this will lay some of the groundwork for that by
setting up Grafana.