Don't validate JSON schemas when running in Docker
The `json-schema` gem is a development and test dependency only. In our
Dockerfile, we only install production gems. Therefore, the
`json-schema` gem is not available when running Docker (even when
running it locally in development mode). Thus, we need this condition to
avoid an error about JSON::Schema being undefined when accessing a path
that would otherwise attempt to validate a JSON response.