[vector] Add fallback for PAPERTRAIL_HOST_AND_PORT
In #5814 , we tried to make it easier to avoid sending logs to
Papertrail by simply commenting out the `PAPERTRAIL_HOST_AND_PORT` env
var.

However, without a fallback/default value, I think that this actually
then makes the Vector configuration invalid, and causes Vector to not do
anything.

I didn't notice this before, because we currently only have the
Papertrail sink. But now I am working on also adding a Loki sink, which
brought to my attention the fact that having a blank
PAPERTRAIL_HOST_AND_PORT not only disables logging to Papertrail, but it
breaks Vector in general.

This change will allow Vector to keep working overall (e.g. sending logs
to Loki, in the future), even if there is no PAPERTRAIL_HOST_AND_PORT
env var set in the environment.