[development docker] Don't send DataMonitorMailer.expectation_not_met
We cannot send emails in development mode when running in Docker,
because in development the email service is letter_opener, which is a
development-only gem, whereas Docker installs only production gems. (I
guess that letter_opener cannot be made to work within Docker, anyway,
since the `worker` service machine does not have a browser?)

(Due to using a seeded database when running Docker in development, many
of the data monitor checks will fail, so we would be trying to send
these emails a decent amount of the time. For example, we have only one
user, rather than the 4..100 expected.)