Merely log emails when running via Docker in development (#5822)
`letter_opener` is not available in Docker (because we only install
production gems in Docker, and `letter_opener` is in the `development`
group). Therefore, when running in development mode in Docker, with this
change, we will instead now use `Mail::LoggerDelivery` (which is
provided by the `mail` gem) as the `delivery_method`, rather than
`letter_opener`.