[roll-out-web] Log results of all curl requests to new web container
Lately, we have been getting deployment failures that seem to be
occurring on the line of code being modified herein. Example:
https://github.com/davidrunger/david_runger/actions/runs/12812774650/job/35725480230

To give better visibility into what might be causing those deployment
failures, this change adds more verbose logging that shows the result of
each `curl` request to the new web container.

Additionally, we herein increase the delay between `curl` requsets from
1 second to 2 second, thereby increasing the overall allowed time from
30 seconds to 60 seconds. Motivation: it's possible that the cause of
the deployment failures is simply that the new container is not becoming
available quickly enough (i.e. within 30 seconds). If that is the case,
then increasing the allowed time should be a straightforward solution.
If that is not the case (which might be revealed or at least hinted at
by this new logging), then we can go back to 30 seconds or even less.