[roll-out-web] Remove 'status=waiting' log lines (#5839)
They are not worth it and they clutter up the output too much.

Also, move the attempt number to beginning of curl result log lines.
This way, it will align with the opening log line about the request that
is about to be made, which also starts with the attempt number.