[Gemfile] Remove 'ransack' (#5841)
`ransack` was added to the `Gemfile` in #5134, so that we could install
it from GitHub. In PR #5471, we stopped installing it from GitHub, and
should have removed it completely from the Gemfile, but didn't. This
change does that. (Note that `ransack` is still included in our bundle
as a dependency of `activeadmin`, though.)