[docker-compose/nginx] Make healthcheck request to env var hostname
As of #5844, we can no longer make successful requests to our production
NGINX that don't have a hostname of davidrunger.com.

Locally, though, we want to check localhost, not davidrunger.com.

Therefore, let's use an environment variable to specify the hostname for
the healthcheck request.

In my local `.env.nginx.local` file, I'll set `HEALTHCHECK_HOSTNAME` to
localhost. In the production file, I'll set it to davidrunger.com.