[nginx] Add dedicated health check endpoint
During deployments lately ([latest example][1]), `nginx` has been
considered unhealthy, which is failing the deploy.

I think that this is because the heavy CPU usage during deployment maybe
sometimes slows down the server response times to the point that the
healthcheck fails.

This change adds a dedicated healthcheck that doesn't hit the Rails
server or DNS / the general Internet, with the hope that it will
therefore be more reliable.

Also, because we are not logging this request and because it is less
costly than the previous healthcheck request (since it no longer hits
the Rails app or uses SSL), I am reducing the healthcheck interval from
30s to 10s.

A nice side benefit of this is that it allows us to get rid of the
HEALTHCHECK_HOSTNAME env var.

[1]: https://github.com/davidrunger/david_runger/actions/runs/12856260739/job/35842770187