Log original client IP from CF-Connecting-IP header [DEV-136] (#5864)
This change uses the `CF-Connecting-IP` header to log the original
client IP address (rather than the IP address of the connecting
Cloudflare machine, as we are currently doing). In the event that no
such header is present, we fall back to the `$remote_addr` value
provided by NGINX.

This header can be trivially faked (as we do [in our own NGINX
healthcheck][1], for example), but in the 99+% of cases where the
request is not that malicious/deceitful, I think it will be helpful.
Certainly, on net, I expect that we will be better off versus just
logging the IP address of some Cloudflare box.

[1]:
https://github.com/davidrunger/david_runger/blob/e4728c6e/docker-compose.yml/#L130