Allow viewing "Publicly viewable" logs without logging in [LOG-28] (#5885)
Motivation: I think that the phrasing "Publicly viewable"
implies/suggests that anyone should be able to view the log, even
without logging in. And, in general, I think that is the more useful
feature, i.e. to be able to share a log that doesn't require logging in
to view it.